refactor: typing

+++ rena/views/register.py
@@ -1,4 +1,4 @@
-from discord import Interaction
+from discord import Interaction, User
 from discord.enums import ButtonStyle
 from discord.ui import View, button
 
@@ -12,8 +12,9 @@ class RegisterView(View):
 
     @button(label="확인", style=ButtonStyle.green, emoji="✅")
     async def confirm(self, _, interaction: Interaction):
-        await self.bot.db.insert("User", (interaction.user.id, "", "", 0, 0, True, True))
-        await interaction.response.edit_message(content="성공적으로 가입 됐어! 레나가 학교생활을 도와줄게!", embed=None, view=None)
+        if isinstance(interaction.user, User):
+            await self.bot.db.insert("User", (interaction.user.id, "", "", 0, 0, True, True))
+            await interaction.response.edit_message(content="성공적으로 가입 됐어! 레나가 학교생활을 도와줄게!", embed=None, view=None)
 
     @button(label="취소", style=ButtonStyle.red, emoji="✖️")
     async def cancel(self, _, interaction: Interaction):
@@ -30,8 +31,9 @@ class UnRegisterView(View):
 
     @button(label="확인", style=ButtonStyle.green, emoji="✅")
     async def confirm(self, _, interaction: Interaction):
-        await self.bot.db.delete("User", interaction.user.id)
-        await interaction.response.edit_message(content="탈퇴됐어... 다음에 또 만나자...!!", embed=None, view=None)
+        if isinstance(interaction.user, User):
+            await self.bot.db.delete("User", interaction.user.id)
+            await interaction.response.edit_message(content="탈퇴됐어... 다음에 또 만나자...!!", embed=None, view=None)
 
     @button(label="취소", style=ButtonStyle.red, emoji="✖️")
     async def cancel(self, _, interaction: Interaction):
